Decompile func_800D249C_E144C

+++ include/functions.us.h
@@ -667,7 +667,7 @@ s16 func_800C7924_D68D4(s16 arg0, s16 arg1, s16 arg2, s16 arg3, s32 arg4, s32 ar
 void func_800C80F0_D70A0(u16 arg0, u16 arg1, u16 arg2, s16 arg3);
 void func_800C8184_D7134(s8 arg0, s8 arg1, s8 arg2, s16 arg3);
 void func_80124118_1330C8(VehicleInstance *arg0, s16 arg1);
-s16 func_800D249C_E144C(s16 arg0, s16 arg1, s16 arg2, s16 arg3, s32 arg4, s32 arg5, s32 arg6, s32 arg7);
+s16 func_800D249C_E144C(s16 arg0, s16 arg1, s16 arg2, s16 arg3, s16 arg4, s16 arg5, s16 arg6, u8 arg7);
 void func_800D25A4_E1554(s16 arg0);
 void func_802DFE68_323FB8();
 void func_802D7548_31B698(void *arg0, s32 arg1);
